+++ src/App.tsx
@@ -18,60 +18,71 @@ import { GlobalStyle } from "./style/global";
 import { theme } from "./style/theme";
 import MyCoach from "./pages/MyCoach";
 import CoachRoutine from "./pages/Routine/CoachRoutine";
-
-const router = createBrowserRouter([
-  {
-    element: <Layout />,
-    children: [
-      {
-        path: "/",
-        element: <Home />
-      },
-      {
-        path: "login",
-        element: <Login />
-      },
-      {
-        path: "signup",
-        element: <Signup />
-      },
-      {
-        path: "mypage",
-        element: <Mypage />
-      },
-      {
-        path: "coach-list",
-        element: <CoachList />
-      },
-      {
-        path: "routine",
-        element: <MyRoutine />
-      },
-      {
-        path: "check-password",
-        element: <CheckPassword />
-      },
-      {
-        path: "record",
-        element: <RecordDatail />
-      },
-      {
-        path: "notification",
-        element: <Notification />
-      },
-      {
-        path: "routine/my-coach",
-        element: <MyCoach />
-      },
-      {
-        path: "routine/my-coach/:coachId",
-        element: <CoachRoutine />
-      }
-    ]
-  }
-]);
+import PrivateLayout from "./components/layout/PrivateLayout";
+import { useState } from "react";
 
 function App() {
+  const [isPasswordConfirmed, setIsPasswordConfirmed] = useState(false);
+  const router = createBrowserRouter([
+    {
+      element: <Layout />,
+      children: [
+        {
+          path: "/",
+          element: <Home />
+        },
+        {
+          path: "login",
+          element: <Login />
+        },
+        {
+          path: "signup",
+          element: <Signup />
+        },
+        {
+          path: "mypage",
+
+          element: (
+            <PrivateLayout isAccess={isPasswordConfirmed}>
+              <Mypage />
+            </PrivateLayout>
+          )
+        },
+        {
+          path: "coach-list",
+          element: <CoachList />
+        },
+        {
+          path: "routine",
+          element: <MyRoutine />
+        },
+        {
+          path: "check-password",
+          element: (
+            <CheckPassword
+              onPasswordConfirmed={() => setIsPasswordConfirmed(true)}
+            />
+          )
+        },
+        {
+          path: "record",
+          element: <RecordDatail />
+        },
+        {
+          path: "notification",
+          element: <Notification />
+        },
+        {
+          path: "routine/my-coach",
+          element: <MyCoach />
+        },
+        {
+          path: "routine/my-coach/:coachId",
+          element: <CoachRoutine />
+        }
+      ]
+    }
+  ]);
   return (
     <QueryClientProvider client={queryClient}>
       <ThemeProvider theme={theme}>
